+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 NCS Navigator Configuration gem history
 =======================================
 
+0.3.1
+-----
+
 0.3.0
 -----